Refactore: - Removed StudyManager in StudyManagement
